debuting "Kaisei Opti"
Kaisei Opti has become a part of the "Aga-chuu" brand since replacing the Hachi Maru Pop (as part of the 2024 redesign) last 2 Apr 2024. this font encapsulates the dynamicity of the newly-inducted Kanji that the vtuber is associated with.

starting 9 Jun 2024, it's been deployed across all of Marias' websites as part of its ongoing effort to serve the audience better.